No.
The last line of the script will not print 139926795932424. Here’s why:

a = [1, 2, 3, 4]: a is assigned a list with these elements.
a = a + [5]: This operation creates a new list by concatenating [5] to the original list [1, 2, 3, 4]. The result is a new list [1, 2, 3, 4, 5], and a is updated to reference this new list.
id(a): The id() function returns the memory address of the object a references.